Reorganize CLI code
